requirements.txt
pillow==9.1.1
python-telegram-bot==13.12